(#4878) Improve efficiency of database query counting nodes.
The Node::count_by_currentness_and_successfulness method was fetching all
the data to Ruby and performing the calculations there, which was slow.
This will now be done with SQL's COUNT.

Whether or not this query should be done is something that needs to be
decided on in a future release.